Remove visibility hints
Turns out, when _everything_ is `inline` in header
files, _everything_ must also be visible in the .so.

This didn't detonate on me until compiling a consumer
of libadt.

I still haven't found good info on if inlining from .a
files is something the compiler is powerful enough to
do, but the -flto docs in `man gcc` seem to suggest that
it's very fragile and specific even between plain .o
files, so probably not.